pyproject.toml
[project]
name = "physics-generator"
version = "0.1.0"
description = "基于多 Agent 编排的 CPhO 物理竞赛题全自动生成与审核系统"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "openai>=1.0.0",
    "python-dotenv>=1.0.0",
    "pydantic>=2.0.0",
    "pyyaml>=6.0",
    "PyMuPDF>=1.24.0",
    "html2text>=2024.2.26",
]

[project.scripts]
physics-generator = "app:_cli"

[tool.uv]
dev-dependencies = [
    "pytest>=8.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = [
    "src/app",
    "src/client",
    "src/config",
    "src/formatter",
    "src/generator",
    "src/graph",
    "src/model",
    "src/prompts",
    "src/reader",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
